requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
google-genai
openai
pillow
pandas
openpyxl
psutil
pytest
